# Split up Graph::boruvka_emulation into smaller pieces