--- a/include/graph.h
+++ b/include/graph.h
@@ -58,18 +58,15 @@ class Graph {
    * @param query  an array of supernode query results
    * @param reps   an array containing node indices for the representative of each supernode
    */
-  void sample_supernodes(
-    std::pair<Edge, SampleSketchRet> *query,
-    std::vector<node_id_t> &reps
-  );
+  void sample_supernodes(std::pair<Edge, SampleSketchRet> *query, std::vector<node_id_t> &reps);
 
   /**
    * @param copy_supernodes  an array to be filled with supernodes
    * @param to_merge         an list of lists of supernodes to be merged
    *
    */
-  void merge_supernodes(Supernode** copy_supernodes, std::vector<node_id_t> new_reps,
-                        std::vector<std::vector<node_id_t>> to_merge, bool first_round,
+  void merge_supernodes(Supernode** copy_supernodes, std::vector<node_id_t> &new_reps,
+                        std::vector<std::vector<node_id_t>> &to_merge, bool first_round,
                         bool make_copy);
 
   /**