--- a/src/graph.cpp
+++ b/src/graph.cpp
@@ -118,7 +118,7 @@ void Graph::batch_update(node_id_t src, const std::vector<size_t> &edges, Supern
   supernodes[src]->apply_delta_update(delta_loc);
 }
 
-void Graph::sample_supernodes(std::pair<Edge, SampleSketchRet> *query,
+inline void Graph::sample_supernodes(std::pair<Edge, SampleSketchRet> *query,
                std::vector<node_id_t> &reps) {
   bool except = false;
   std::exception_ptr err;
@@ -137,7 +137,7 @@ void Graph::sample_supernodes(std::pair<Edge, SampleSketchRet> *query,
   if (except) std::rethrow_exception(err);
 }
 
-std::vector<std::vector<node_id_t>> Graph::supernodes_to_merge(std::pair<Edge, SampleSketchRet>
+inline std::vector<std::vector<node_id_t>> Graph::supernodes_to_merge(std::pair<Edge, SampleSketchRet>
                *query, std::vector<node_id_t> &reps) {
   node_id_t size[num_nodes];
   std::fill(size, size + num_nodes, 1);
@@ -197,7 +197,7 @@ std::vector<std::vector<node_id_t>> Graph::supernodes_to_merge(std::pair<Edge, S
   return to_merge;
 }
 
-void Graph::merge_supernodes(Supernode** copy_supernodes, std::vector<node_id_t> new_reps,
+inline void Graph::merge_supernodes(Supernode** copy_supernodes, std::vector<node_id_t> new_reps,
                std::vector<std::vector<node_id_t>> to_merge, bool make_copy, bool first_round) {
   bool except = false;
   std::exception_ptr err;