I moved the code in Graph::boruvka_emulation for sampling supernodes, identifying outgoing edges, and merging supernodes to their own functions.
To make this easier, I promoted the local "modified" variable in the do/while loop to a member variable of the Graph class. This way, Graph::supernodes_to_merge can return just one value.