@etwest 2022-02-21
> Do you have numbers on how these changes effect query performance?

Do you have access to our lab machine? If not I can run this experiment to get these numbers.