@edeturk32 2022-02-22
> Do you have access to our lab machine? If not I can run this experiment to get these numbers.
I don't think I have access to the lab machine