@etwest 2022-02-22
> I don't think I have access to the lab machine

Okay I will run the experiment later today. I'll post the results here after that and we can go from there.