@etwest 2022-02-23
| Backup In Mem | Branch: main, cc alg latency | Branch: cc_split, cc alg latency |
|-|-|-|
| ON | 1.25, 1.28, 0.40, 0.32, 0.32, 0.32, 0.32, 0.32, 0.32, 0.32 | 1.40 1.24 0.51 0.33 0.33 0.33 0.33 0.34 0.33 0.33 |
| OFF | 23.6, 23.6, 24.0, 24.0, 25.0, 25.6, 25.6, 25.5, 25.6, 25.7 | 26.5, 26.0, 26.0, 25.9, 26.0, 26.1, 26.0, 25.6, 25.8, 25.8  |

Looks like your changes make the cc algorithm a little bit slower.